fixed all warnings again, used size_t where obvious
git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@3957 0e98fd9d-079e-457f-9c5c-d482923f975d

+++ src/marsyas/ExNode.cpp
@@ -182,7 +182,7 @@ ExNode::is_list() const
 {
 	std::string humuhumunukunukuapuaa=getType();
 	// whoa! that's crazy man..
-	uint32_t len=(uint32_t)humuhumunukunukuapuaa.length();
+	size_t len = humuhumunukunukuapuaa.length();
 	return (len>3)
 		&& (humuhumunukunukuapuaa[len-4]=='l')
 		&& (humuhumunukunukuapuaa[len-3]=='i')
